// src/app/disclosure-dialog/disclosure-dialog-interface.ts
export interface messagesDialog {
  title: string;
  btnLeftDisable: boolean;
  btnRightDisable: boolean;
  txtBtnLeft: string;
  txtBtnRight: string;
  HEADER_1_TH: string;
  HEADER_2_TH: string;
  DETAIL_1_OF_HEADER_1_TH: string;
  DETAIL_2_OF_HEADER_1_TH: string;
  DETAIL_3_OF_HEADER_1_TH: string;
  DETAIL_1_OF_HEADER_2_TH: string;
  DETAIL_2_OF_HEADER_2_TH: string;
  DETAIL_3_OF_HEADER_2_TH: string;
  DETAIL_4_OF_HEADER_2_TH: string;
  DETAIL_5_OF_HEADER_2_TH: string;
  DETAIL_6_OF_HEADER_2_TH: string;
  DETAIL_7_OF_HEADER_2_TH: string;
  CONSENT_TO_DISCLOSURE_TH: string;
  HEADER_1_EN: string;
  HEADER_2_EN: string;
  DETAIL_1_OF_HEADER_1_EN: string;
  DETAIL_2_OF_HEADER_1_EN: string;
  DETAIL_3_OF_HEADER_1_EN: string;
  DETAIL_1_OF_HEADER_2_EN: string;
  DETAIL_2_OF_HEADER_2_EN: string;
  DETAIL_3_OF_HEADER_2_EN: string;
  DETAIL_4_OF_HEADER_2_EN: string;
  DETAIL_5_OF_HEADER_2_EN: string;
  DETAIL_6_OF_HEADER_2_EN: string;
  DETAIL_7_OF_HEADER_2_EN: string;
  CONSENT_TO_DISCLOSURE_EN: string;
  CONFIRM_TH: string;
  CONFIRM_EN: string;
}